fix(viewer): split CommandPalette's search/ranking helpers into their own module
CommandPalette.tsx grew to 852 lines on main against its 849-line
module-size budget after #3942 and #3945 each landed a command entry the
same day, so check-module-size.mjs exits 1 on a clean checkout and every
open PR inherits the red.

Extract the pure fuzzy-search/ranking and recent-usage-tracking helpers
(Category/Command/FlatItem types, score/rankCommand, getRecentIds/
recordUsage) into commandPaletteSearch.ts — none of it touches React or
the store, a clean seam. CommandPalette.tsx is now 752 lines (97 lines
of headroom); every command id is unchanged and the new module is
imported and used in place.

Closes #3957
